Not ready on Non-Windows platform. (#178)
Not ready on Non-Windows platform for **.NET 3.0 Preview5**. This is related to this bug fix: https://github.com/dotnet/coreclr/pull/24198, which will be included in **Preview6**.
The issue manifests when EventPipe is enabled without LTTng. In this case, GC events are disabled after the first GC because LTTng probes are disabled.