Improve perf of `!dumpasync -tasks` (#294)
When the `-tasks` flag is set, we look at every object on the heap and walk its hierarchy to decide if it's a task or a task-derived type.  But we currently do that by comparing the name of every object against "System.Threading.Tasks.Task", which is very expensive.  This switches to comparing by module and type def.  On my machine, a `!dumpasync -tasks` command that previously took ~1 minute now takes ~30 seconds.