Fix missing metadata problems
The previous fix forced the metadata to always go through
the metadata locator callback interface. This fix the problem
when the metadata in the core dump isn't complete, but broke
getting metadata for user/app assemblies that only exist in
the core dump.

Needed to "fallback" to reading the core dump memory for the
metadata when the assembly can't be found or downloaded.

Cleaned up building the module list.