Add Microsoft.DiaSymReader.Native to SOS package
We no longer search for the diasymreader in the runtime host
directory, DAC path, etc. because for WER/Watson/!analyze needs
to work without assuming there is a .NET Core runtime to host
on the machine.

Update to version 1.1.35504 of clrmd. Fixes issues https://github.com/dotnet/diagnostics/issues/360
and "System.DllNotFoundException: Unable to load shared library 'libdl.so'..." failure.