interleave IL in disassembly with some code dup
The current implementation successfully interleaves the
MSIL (Microsoft Intermediate Language) along with disassembly
via the "!u -il <IP>" command in the windbg command window.
This is a first cut implementation with reasonable code
duplication. There is a scope for code cleanup but the
current implementation should work out for now.