working workaround issues in map with 2 entries
Note: The previous commit did not fix the issue.

There is an issue with respect to the map generate by the debugger,
wherein the same intermediate language has two entries (can be
multiple as well). One of those entries has start address
same as that of the end address. This entry can be safely ignored
and this is exactly what this commit attempts to achieve.
As an example see the following:

Looks like the following IL is missing and needs to be debugged, which
is strange because it should be a simpler case to deal with.

```
IL_000c: call string System.Console::ReadLine()
```

The following is the source of issue which either must be fixed in the
map generated by debugger or any workaround towards this issue.

```
000c 00007FFD03BC0F99 00007FFD03BC0F99
000c 00007FFD03BC0F99 00007FFD03BC0FA2
```