Build Windows arm64 version of SOS.
Make sure that the "BuildArch" msbuild property is always passed to managed and native builds.

Add the arm64 SOS to the various packages and zip files.

The diasymreader package for Windows PDB support is currently disabled for arm64 waiting
for issue https://github.com/dotnet/diagnostics/issues/324 to be addressed.