Change the "dotnet-dump collect" default to "full" dumps.
The new "full" dump collection type includes all the module images
on both Windows and Linux. The old default type "Heap" is still
the same and can be used if the dump is too big.

Full dumps are roughly 1.5 times bigger on Windows and 30%
bigger on Linux for a simple webapp.

Issue: https://github.com/dotnet/diagnostics/issues/808