Add .NET Core hosting location file support (#915)
* Add .NET Core hosting location file support

Look for the /etc/dotnet/install_location file on Linux/MacOS
for the .NET Core runtime to host SOS's managed code.

Didn't add the check for Windows registry key at this time
because we may switch to the desktop CLR as the host.

Issue #https://github.com/dotnet/diagnostics/issues/882

Add 5.0 to the list of hosting runtime versions.

* Code review feedback