Fix the dumpil /i on Linux
On Linux/MacOS '/' isn't allowed as an option but the dumpil command had '/i'. The reason '/' is not allowed on xplat is that it caused problems parsing file paths as args. It most cases '-' is used except for this case and for the dml '/d' which isn't supported on lldb.

I'll fix it by adding -i as an option to dumpil and add it to the doc.

This also give me the chance to remove a bunch of unnecessary #ifndef FEATURE_PAL around the "/d" DML option.

Issue #https://github.com/dotnet/diagnostics/issues/1478