Fixes for some vendors issues for the stable release (#1785)
* Fixes for some vendors issues for the stable release

1) Stop printing the "@(#)Version" prefix of the SOS version on Linux.
2) Change the default symbol cache for dotnet-dump on Windows to VS's.
3) Fix eeversion unhandled exception on Linux when hosted on 5.0.0.
4) Fix `dotnet-dump collect --help` ArgumentOutOfRangeException by reformatting the --type option description.

* Fix eeversion unhandled exception on Linux hosted on 5.0.0

This turned out to be that the dotnet-sos installed included runtime binaries
like System.Memory, System.Buffer, etc. that were wrong version for hosting on
5.0.0. This didn't get catch by the normal testing because only certain project
references were copied to the bin directory.

The main part of this fix was to be very explicit in which assemblies are added to
the TPA that are in the SOS directory instead of adding all of them.

Changed the SOS managed installed to copy all the managed assemblies in SOS.NETCore
to better match the dotnet-sos install layout.

Changed the SOS tests to explicitly test across all the host runtimes (2.1.x,
3.1.x, 5.0.x and desktop).

* Fix `dotnet-dump collect --help` ArgumentOutOfRangeException by reformatting the --type option description.