SOS/dotnet-dump extensibility - Part 1 (#1823)
* Split the interface and implementation debug service assemblies

* Add IHost and ITarget interfaces.  Add the Target and Target data reader implementations.

Add ISymbolService and an implementation that temporarily uses SOS.NETCore.

ServiceProvider cleanup.

SOS.Hosting cleanup.

* Add ISymbolService and implementation temporarily using SOS.NETCore.

* Add IMemoryService and implementations

There are two helper IMemoryService implementations PEImageMappingMemoryService and MetadataMappingMemoryService. They are used to wrap the base native debugger memory service implementation.

PEImageMappingMemoryService is used in dotnet-dump for Windows targets to mapping PE images like coreclr.dll into the memory address space the module's memory isn't present. It downloads and loads the actual module and performs the necessary fix ups.

MetadataMappingMemoryService is only used for core dumps when running under lldb to map the managed assemblies metadata into address space. This is needed because the way lldb returns zero's for invalid memory for dumps generated with createdump on older runtimes (< 5.0).

The address sign extension plan for 32 bit processors (arm32/x86) is that address are masked on entry to the managed infrastructure from DAC or DBI callbacks or from the native SOS code in SOS.Hosting. If the native debugger that the infrastructure is hosted needs addresses to be signed extended like dbgeng, it will happen in the debugger services layer (IDebuggerService).

* Add IModuleService and IModule interfaces and implementation

* Update IThreadService, add IThread (replacing ThreadInfo) and add new ThreadService implementation.

* Add IRuntimeService/IRuntime interfaces and implementations.

This service abstracts the .NET runtimes in the process.

* Update IConsoleService and add ICommandService and update the implementatiion.

Command invocation is now synchronous so it will work under dbgeng.

Cleaned up the command and option alias attributes.

* Update test scripts.