Make stresslog work for clrgc.dll, increment SOS_BREAKING_CHANGE_VERSION (#2272)
We would like to run tests with clrgc.dll activated and compiled with USE_REGIONS - as it turns out, the stresslog didn't yet work with the in-memory implementation of the stresslog. 

This has the changes required so we are actually able to dump this version of the stresslog as well. It also increments SOS_BREAKING_CHANGE_VERSION because a change in the StressLog data structure was required to support logging from clrgc.dll.