Fix unknown type/methods in core dumps.
Fixes some of the issues in https://github.com/dotnet/diagnostics/issues/2375

The problem is that the image mapping memory service didn't convert the rva from a
loaded layout calculated from the in-memory module to the file layout (the PEReader
with the downloaded image).

On Windows, images (native or managed) are always loaded layout so return false in
IModule.IsFileLayout without calling GetPEInfo() to avoid the recursion that broken
getting the info about coreclr.dll. It turns out that the heap dumps generated on
Windows don't have the image in-memory.

Don't get module version in GetPEInfo() to determine the layout. Cleanup.

Skip relocations that span cache blocks. This happens very rarely and should not affect
anything unless we get really really unlucky.