Support Windows/MacOS single-file in the C++ runtime enum fallback
IDebuggerServices::GetOffsetBySymbol is used to get the runtime info export and so most
of the changes are making the interface available on the libsos "side" under Linux/MacOS
since there is a Extensions instance both in libsosplugin and libsos. The libsos side
Extensions::GetDebuggerServices always returned null before this change. Now it only
returns null if hosted under a managed host like dotnet-dump.