Enhance core dump debugging
I've found that when a core dump is taken by the OS without the extra
settings that makes it copy shared libraries binaries into the dump,
SOS doesn't work as it relies on the libcoreclr.so ELF header to be
present.
I have discovered a way to make lldb read that data from the image of
the libcoreclr.so that it has loaded together with the dump.

This change makes it possible to use SOS even for the above mentioned
dump cases. In fact, it allows SOS to read any data that were mapped as
read only from any shared library loaded in the process.