Remove stale gauges from dotnet-counters UI
dotnet-counters currently adds new metrics and tags that it discovers to the console
UI but never removes anything, even if the data for those metrics/tags stops coming.
For gauges the lack of data is meaningful because the user code explicitly chose not
to provide a value when queried. This change removes metrics/tags from the UI if the
gauge in the monitored process stops sending data.