Missing memory improvements (#3662)
* Improve native memory commands when missing memory

I made a few improvements from using this on a couple of investigations.

We now print a warning when we notice that regions we know exist in memory are missing from IMemoryRegionService.  This shouldn't happen on Windows, but in ELF dumps we do not have a debug stream describing memory...we only use what was put into the coredump as the known memory regions.

We now split up VirtualAlloc regions when we go to tag CLR memory regions.  CLR might VirtualAlloc one big chunk and sub-divide it into its own regions.  Note that MAddress isn't actually trying to show VirtualAlloc/memmap addresses (just use !address for that), instead it's trying to take what !address knows and annotate more information.

* Add [] around MEM_RESERVE and MEM_FREE to better distinguish them

* Remove Index

Remove some dead code.

* Documentation edits