Add !dumpheap fragmentation statistics (#3799)
In my previous !dumpheap change I overlooked the fragmentation output.  This adds fragmentation output in the exact same was as the previous C++ code.

The new code now validates the Free region is actually followed by the next object, and that those objects do not live on the Pinned, Frozen, or Large object heaps.