Check for cancellation when reading memory
DbgEng returns a specific error code when the user hits Control-C.  We can work around any command or code which forgets to check for cancellation by throwing this manually when we detect the cancellation error code.

This should also improve ClrMD's handling of Control-C.  A read that _should_ complete (but failed due to cancellation) may cause ClrMD to cache the fact that the read failed.  By throwing an exception we will bypass any place where we would have interpreted that failed read as something other than cancellation.