Add static variable names to !gcroot
Static variables can be opaque to the user when looking at !gcroot, since they show up as a pinned handle to an object array.  It's also difficult to distinguish things _other_ than static variables which follow the same pattern.  This change adds the static variable name to the output of !gcroot to help distinguish these differences.

- Remove "this will take a while" warning output from !dumpheap -short.
- Added StaticVariableService to resolve the address/objects of static variables.  This can be a relatively expensive operation, so it's important to cache the result.
- Updated GCRootCommand with the heuristic.