Add generation option to dumpheap command (#3966)
This is a very small change to address https://github.com/dotnet/diagnostics/issues/3963

Few questions:
1) Should I add tests for this?
2) I've asked about GCGeneration custom type that is a clone of
Generation "official" type in that issue. It is a part of dumpgen
command which itself, after this change, kinda loses it's meaning. I
think the way forward is just to either remove dumpgen command or leave
it as a proxy for dumpheap.