!maddress misattributes memory to the GC (#4318)
Fixes an issue reported by the GC team.

On windows, VirtualQuery information (bounds of allocated segments) is
very accurate in dumps because we have an entire segment of the dump
format which describes it. On Linux, it seems like we smash together
regions which came from two different calls of VirtualAlloc (well, the
Linux equivalent).

So, only use a heuristic of attributing regions of allocated memory on
Windows. Additionally, we are 100% accurate in reporting every byte of
memory that the GC allocates as of .Net 8. We shouldn't be ever using
this heuristic to attribute any memory to the GC.