Fix log scope construction (#4323)
This PR reworks how log scopes are reconstructed. The current
implementation has a several issues that results in it only working in
single-threaded contexts. Once multiple threads are involved, scopes
will be incorrectly associated to log messages in addition to scopes
being deleted early while they are still active.

I've described how the new approach works in the source code so it
doesn't get lost after this PR is merged. I've also left a comment with
a visual representation of edge cases with concurrent traces.

/cc @dotnet/dotnet-monitor

---------

Co-authored-by: Wiktor Kopec <wiktork@microsoft.com>