Add dotnet-counters console exporter tests
Today all the testing of the console formatting is mostly manual.
This adds some automated testing of the formatting logic in preparation
to both refactor and make some design changes. The tests should make
it very obvious in the future when the output formatting intentionally
changes.

This change also includes a small update to opportunistically hide
the cursor when doing dotnet-counters monitor which looks a little
nicer.