Refactor _maxNameLength
This variable was being used inconsistently in the code. The correct
meaning was max allowable name length, not largest name so far.
There was no bug before because any time the code treated it
as max name length, later the code called AssignRowsAndInitializeDisplay
which reset the variable back to its correct value. It is confusing
though to see code temporarily set the variable to an incorrect and
unused value before later setting it correctly.

# Conflicts:
#	src/Tools/dotnet-counters/Exporters/ConsoleWriter.cs