Counter instruments are absolute values now
Fixes #3751

Previously we always showed counter instruments as a rate, but
for a variety of counters that isn't the most useful value. Now
we are showing the absolute value as long as the target process
is running .NET 8. For cases where it was useful to see rate of change
there is now a --showDeltas option which enables a 2nd column
of output for dotnet-counters monitor. The deltas show the difference
between the previous value and the current value. Making the new
column optional rather than a default setting is intended to minimize
consuming horizontal line space which is already at a premium. When
using csv or json outputs it is assume the user's tool that processes
the data later can compute the rate if desired so there is no special
delta handling added for those.

Take a look at the updated exporter tests to see the specific changes
in dotnet-counters output format.

For other tools consuming System.Diagnostics.Monitoring.EventPipe
there is a new setting on the settings object that allows opting in to
the new behavior. I'm guessing dotnet-monitor will want to do it too
but I wanted to ensure they can opt-into the new behavior at a
time of their choosing.