Add support for new exception trace storage format
To eliminate global spinlock in stack trace saving to exceptions, the
format of the stack trace storage has changed. The `_stackTrace` array
can now be either a byte[] with the actual stack trace as before or an
object[] where the first element references the byte[] with the actual
stack trace and the following elements contain references to
System.Resolver or System.Reflection.LoaderAllocator instances that keep
code of methods that can be collected alive.

This change modifies SOS to handle the object[] case properly, while
staying compatible with the old way. Since the same code to get the
stack trace was duplicated at two places, I have taken this opportunity
to refactor it into a separate function.