Allow SOS testing against a private runtime build (#4818)
Added eng\privatebuild.cmd and eng/privatebuild.sh and some build
changes that sets up for testing against a private build. Just the
latest 9.0.x runtimes are installed and tested against.

Added eng/testsos* as a short cut to running just the SOS tests.

Updated the doc on the end-to-end procedure.