Metric name disambiguation (#4938)
In .NET 8.0 we added tags on Meters and Instruments but
MetricsEventSource only included the names when emitting value
publishing events. This made it ambiguous when there was more than one
Meter or Instrument with the same name. In 9.0 MetricsEventSource
started included InstrumentIDs in the BeginInstrumentReporting events
and in the value publishing events that provides a stronger correlation
identifier. This change consumes the IDs in those events within
Microsoft.Diagnostics.Monitoring.EventPipe, allowing dotnet-counters
(and presumably dotnet-monitor too) to track time series that differ
only by the tags provided to the Meter/Instrument constructors.

I also re-enabled a disabled test that got broken when .NET 9.0 added
the System.Runtime Meter.

Fixes #4843, #4564