Merged PR 43913: Add DAC cert check
Add DAC cert check

Calls the CLRMD signing and cert checking helper function before loading the DAC module.

The check honors the Debugger.Settings.EngineInitialization.SecureLoadDotNetExtensions setting when running under windbg. Defaults to true under dotnet-dump on Windows, false on Linux/MacOS for dotnet-dump and lldb.

TestHost (used by the Microsoft.Diagnostics.DebugServices.UnitTests, etc.) disables the signature check because all the test asset dumps used are preview 6.0 and not signed (that needs to be fixed but it is beyond the scope of this PR).
----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a new feature to check DAC certificate validity.
- Added `GetSettings` method in `dbgengservices.cpp` to retrieve debugger settings.
- Implemented `DacSignatureVerificationEnabled` property in `HostServices.cs` and `DebuggerServices.cs` to enforce DAC certificate validation.
- Created new interface `ISettingsService` in `ISettingsService.cs` to define settings service.
- Updated `RuntimeWrapper.cs` to verify DAC signing and certificate before loading.
- Modified `Versions.props` to update `MicrosoftDiagnosticsRuntimeVersion` to `4.0.0-beta.24521.1`.